Extract Alias navigation in a separate package
