+++ client/src/views/dTypeLandingPage.vue
@@ -92,7 +92,7 @@ export default {
         if (this.name) {
           hash = await this.contract.getTypeHash(this.lang, this.name);
         }
-        this.typeStruct = await this.$store.dispatch('getTypeStruct', {hash});
+        this.typeStruct = await this.$store.dispatch('dtype/getTypeStruct', {hash});
         this.typeContract = null;
         this.items = [];
         this.headers = this.getHeaders(this.typeStruct);
@@ -106,7 +106,7 @@ export default {
           this.typeContract = await getContract(
             this.typeStruct.contractAddress,
             abi,
-            this.$store.state.wallet,
+            this.$store.state.dtype.wallet,
           );
 
           getDataItems(this.typeContract, (dataItem) => {
@@ -152,7 +152,7 @@ export default {
     },
     insert(dtype) {
       if (this.isRoot) {
-        this.$store.dispatch('insertType', dtype);
+        this.$store.dispatch('dtype/insertType', dtype);
       } else {
         this.typeContract.insert(dtype)
           .then(tx => tx.wait(2))
@@ -161,7 +161,7 @@ export default {
     },
     batchInsert(dtypeArray) {
       if (this.isRoot) {
-        this.$store.dispatch('insertBatchType', dtypeArray);
+        this.$store.dispatch('dtype/insertBatchType', dtypeArray);
       } else {
         for (let i = 0; i < dtypeArray.length; i++) {
           this.insert(dtypeArray[i]);
@@ -179,7 +179,7 @@ export default {
     },
     remove(dtype) {
       if (this.isRoot) {
-        this.$store.dispatch('removeType', dtype);
+        this.$store.dispatch('dtype/removeType', dtype);
       } else {
         this.typeContract.remove(dtype.typeHash)
           .then(tx => tx.wait(2))
