+++ packages/dtype-default-ui/package.json
@@ -8,16 +8,18 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "vue-cli-service lint"
   },
-  "dependencies": {
-    "vue": "^2.6.6",
-    "vuetify": "^1.5.5"
-  },
   "files": [
     "dist/*",
     "src/*",
     "*.json",
     "*.js"
   ],
+  "dependencies": {
+    "vuetify": "^1.5.5"
+  },
+  "peerDependencies": {
+    "vue": "^2.6.6"
+  },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^3.5.0",
     "@vue/cli-plugin-eslint": "^3.5.0",
