+++ packages/dtype-navigation-ui/package.json
@@ -1,17 +1,16 @@
 {
-  "name": "@dtype/markdown-ui",
+  "name": "@dtype/navigation-ui",
   "version": "0.0.1",
-  "description": "dType markdown type",
+  "description": "dType navigation",
   "main": "./index.js",
   "scripts": {
-    "build": "vue-cli-service build --target lib --name dtype-markdown-ui ./index.js",
+    "build": "vue-cli-service build --target lib --name dtype-ui-navigation ./index.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "marked": "^0.7.0",
-    "vue-simplemde": "^1.0.0",
-    "vuetify": "^1.5.5"
+    "@dtype/alias": "0.0.1",
+    "@dtype/core": "0.0.1"
   },
   "files": [
     "dist/*",
@@ -20,9 +19,12 @@
     "*.js"
   ],
   "peerDependencies": {
-    "vue": "^2.6.6"
+    "vue": "^2.6.6",
+    "vuetify": "^1.5.5",
+    "vuex": "^3.1.0"
   },
   "devDependencies": {
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@vue/cli-plugin-babel": "^3.5.0",
     "@vue/cli-plugin-eslint": "^3.5.0",
     "@vue/cli-service": "^3.5.0",
